requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra dev -o requirements-dev.txt pyproject.toml
asgiref==3.8.1
    # via
    #   django
    #   django-browser-reload
    #   django-htmx
black==24.4.0
certifi==2024.2.2
    # via django-tailwind-cli
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via
    #   black
    #   django-typer
    #   djlint
    #   typer-slim
colorama==0.4.6
    # via djlint
crispy-tailwind==1.0.3
cssbeautifier==1.15.1
    # via djlint
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
distlib==0.3.8
    # via virtualenv
django==5.0.4
    # via
    #   crispy-tailwind
    #   django-allauth
    #   django-browser-reload
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-htmx
    #   django-import-export
    #   django-tailwind-cli
    #   django-template-partials
    #   django-typer
    #   djangorestframework
    #   heroicons
django-allauth==0.62.1
django-browser-reload==1.12.1
django-crispy-forms==2.1
    # via crispy-tailwind
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-htmx==1.17.3
django-import-export==3.3.8
django-tailwind-cli==2.11.0
django-template-partials==24.2
django-typer==1.1.2
    # via django-tailwind-cli
djangorestframework==3.15.1
djlint==1.34.1
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
et-xmlfile==1.1.0
    # via openpyxl
filelock==3.13.4
    # via virtualenv
gunicorn==21.2.0
heroicons==2.6.0
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
identify==2.5.36
    # via pre-commit
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.25
    # via djlint
markuppy==1.14
    # via tablib
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==24.0
    # via
    #   black
    #   gunicorn
pandas==2.2.2
pathspec==0.12.1
    # via
    #   black
    #   djlint
platformdirs==4.2.1
    # via
    #   black
    #   virtualenv
pre-commit==3.7.0
psycopg==3.1.18
psycopg-binary==3.1.18
    # via psycopg
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
pywatchman==2.0.0
pyyaml==6.0.1
    # via
    #   djlint
    #   pre-commit
    #   tablib
regex==2023.12.25
    # via djlint
ruff==0.4.1
setuptools==69.5.1
    # via nodeenv
shellingham==1.5.4
    # via django-typer
six==1.16.0
    # via
    #   cssbeautifier
    #   jsbeautifier
    #   python-dateutil
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
tablib==3.5.0
    # via django-import-export
tqdm==4.66.2
    # via djlint
typer-slim==0.12.3
    # via django-typer
typing-extensions==4.11.0
    # via
    #   psycopg
    #   typer-slim
tzdata==2024.1
    # via pandas
virtualenv==20.26.0
    # via pre-commit
whitenoise==6.6.0
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
